image issues, comments, +about.txt, +test file, size function
Fixed the regions of the image capture to be more accurate on the crosshair and above the crosshair. I added a lot more comments. created an about text file for me to note and give more info. Created a test file to run test code that will be implemented in the vl file. Fixed the size function by adding and ___ >= 0. this made it so the file size differences are limited only to between its number and 0. It also makes it so the file size differences can be 0.